Move drawing functions into single file
