--- src/client/canvasctx.js
@@ -1,5 +1,0 @@
-import CANVAS from "./canvas";
-
-const CANVAS_CTX = CANVAS.getContext("2d");
-
-export default CANVAS_CTX;
