+++ src/client/gameloop.js
@@ -1,5 +1,5 @@
 import REQUEST_ANIMATION_FRAME from "./raf";
-import DRAW_RECT from "./drawrect";
+import { DRAW_RECT } from "./draw";
 import Timer from "./timer";
 import GameCanvas from "./gamecanvas";
 import Camera from "./camera";
