+++ src/client/uicommon.js
@@ -1,4 +1,4 @@
-import DRAW_IMAGE from "./drawimage";
+import { DRAW_IMAGE } from "./draw";
 import GameCanvas from "./gamecanvas";
 import WZManager from "./wzmanager";
 import PLAY_AUDIO from "./playaudio";
