+++ src/client/uilogin.js
@@ -1,6 +1,4 @@
-import DRAW_IMAGE from "./drawimage";
-import DRAW_TEXT from "./drawtext";
-import DRAW_RECT from "./drawrect";
+import { DRAW_IMAGE, DRAW_TEXT } from "./draw";
 import GameCanvas from "./gamecanvas";
 import WZManager from "./wzmanager";
 import UICommon from "./uicommon";
